+++ tsg/sampling/point_sampling.py
@@ -4,7 +4,7 @@ from numpy.typing import NDArray
 
 def sample_points(
     num_points: int,
-) -> tuple[NDArray[np.float64], tuple[np.float64, np.float64], np.float64]:
+) -> tuple[NDArray[np.float64], tuple[float, float], float]:
     coordinates = sample_spherical(num_points)
     shift = move_points(coordinates)
     border_values = (
@@ -20,18 +20,16 @@ def sample_spherical(num_points: int, ndim=3) -> NDArray[np.float64]:
     return vec
 
 
-def get_border_value(
-    coordinates: NDArray[np.float64], is_min: bool = True
-) -> np.float64:
+def get_border_value(coordinates: NDArray[np.float64], is_min: bool = True) -> float:
     if is_min:
-        return np.float64(np.min(coordinates))
+        return float(np.min(coordinates))
     else:
-        return np.float64(np.max(coordinates))
+        return float(np.max(coordinates))
 
 
-def move_points(coordinates: NDArray[np.float64]) -> np.float64:
+def move_points(coordinates: NDArray[np.float64]) -> float:
     min_coordinate = get_border_value(coordinates)
-    shift = np.float64(0)
+    shift = 0.0
     if min_coordinate < 0:
         coordinates += np.abs(min_coordinate) * 2
         shift = np.abs(min_coordinate) * 2
